Add files via upload
This Godot Project is all about Teeworlds' movement and it's behavior to the gravity and air force. It has a decent jump mechanic too which is supposedly customizable.
It is free to use for everybody, I don't own Teeworlds but since it's open source, I wanted to replicate the movement the most. Hope it's helpful to you in any way.
It contains:
- Kinematic Body Function
- Velocity based movement with Move and Slide Function
- A lot of descriptions to understand what some calculations were for
- An optional UI for debugging (Currently with 4 specs: X velocity, Y velocity, FPS and jump counter)
-Basic tilemap